Coerce to a Data Frame

Creates data frames from censored and water-quality data: methods for "lcens," "mcens," and "qw" data.

# S3 method for lcens
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  nm = deparse(substitute(x))
)

# S3 method for mcens
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  nm = deparse(substitute(x))
)

# S3 method for qw
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  nm = deparse(substitute(x)),
  expand = FALSE
)

Arguments

x
any R object to put into a data frame.

row.names
NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional
logical. If TRUE, set column names to nm.

...
additional arguments to be passed to or from methods.

nm
set as column name if optional is TRUE.

expand
create a data frame from all of the slots from the object instead of creating a data frame containting the object.

Value

A data frame constructred from x.

See also