Probability Plotting Positions

Generates the sequence of probability points for uncensored and censored values

censpp(x, a = 0.44)

# S3 method for default
censpp(x, a = 0.44)

# S3 method for lcens
censpp(x, a = 0.44)

# S3 method for mcens
censpp(x, a = 0.44)

Arguments

x
a vector of observations

a
the offset fraction to be used; typically in [0, 0.5].

Value

A list with these components:

Note that xrcn and pprcn are provided only if there are right-censored values.

x

sorted uncensored observations.

pp

corresponding plotting positions for uncensored observations.

xcen

sorted left-censored observations.

ppcen

corresponding plotting positions for left-censored observations.

xrcn

sorted right-censored observations.

pprcn

corresponding plotting positions for right-censored observations.

Note

The left-censored plotting positions are adjusted by the method of Helsel and Cohn (1988), which produces a mean of 0.5 for all data.

References

Helsel, D.R. and Cohn, T.A., 1988, Estimation of descriptive statistics for multiply censored water quality data: Water Resources Research v. 24, n. 12, pp.1997-2004

See also

Examples

# This example demonstrates the probability adjustments made for censoring
# The raw, uncensored data:
ppoints(7, .44)

#> [1] 0.07865169 0.21910112 0.35955056 0.50000000 0.64044944 0.78089888 0.92134831

# censor ar 0.5 (the lowest 3 values)
censpp(as.lcens(ppoints(7, .44), 0.5))

#> $x
#> [1] 0.5000000 0.6404494 0.7808989 0.9213483
#> 
#> $pp
#> [1] 0.5062413 0.6449376 0.7836338 0.9223301
#> 
#> $xcen
#> [1] 0.5 0.5 0.5
#> 
#> $ppcen
#> [1] 0.07692308 0.21428571 0.35164835
#>