merge flow variable into analysis data frame and update iSpec with variable name

merge flow variable into analysis data frame and update iSpec with variable name

.mergeFlow(
  ct1 = ct1,
  iSpec = iSpec,
  gageID = gageID,
  hydro.var = hydro.var,
  flow.detrended = flow.detrended
)

Arguments

ct1

analysis data frame

iSpec

iSpec

gageID

"q" + USGS gage ID

hydro.var

averaging windows

flow.detrended

data generated by detrended.flow. Default = flow.detrended.