Plot censored gam fits vs. time

Plot censored gam fits vs. time

gamPlotDisp(gamResult = gamResult, analySpec = analySpec,
  fullModel = 2, seasAvgModel = 2, seasonalModel = 2,
  diffType = "regular", obserPlot = TRUE, interventionPlot = TRUE,
  seasAvgPlot = TRUE, seasAvgConfIntPlot = TRUE,
  seasAvgSigPlot = TRUE, fullModelPlot = TRUE, seasModelPlot = TRUE,
  BaseCurrentMeanPlot = TRUE, adjustedPlot = FALSE)

Arguments

gamResult
output from procedure gamTest

analySpec
analytical specifications

fullModel
GAM # for displaying full GAM (e.g., 0, 1, 2)

seasAvgModel
GAM # for displaying seasonally average GAM

seasonalModel
GAM # for displaying seasonal GAM

diffType
plot predicted baseline mean ('regular') or adjusted baseline mean ('adjusted')

obserPlot
logical field indicating whether to plot observations

interventionPlot
logical field indicating whether to plot interventions (e.g., method changes)

seasAvgPlot
logical field indicating whether to plot seasonal average GAM

seasAvgConfIntPlot
logical field indicating whether to plot confidence interval for seasonal average GAM

seasAvgSigPlot
logical field indicating whether to plot significant increasing and decreasing trends for seasonal average GAM

fullModelPlot
logical field indicating whether to plot full GAM

seasModelPlot
logical field indicating whether to plot seasonal GAM

BaseCurrentMeanPlot
logical field indicating whether to plot baseline and current mean

adjustedPlot
logical field indicating whether to plot adjusted model