Impute Censored Values in dataframes

Impute value for multiply censored data in data frames

imputeDF(df, imputeOption = "mid")

Arguments

df
dataframe

imputeOption
imputation method [default= "mid"], valid impute options are "lower", "upper", "mid", "norm", "lnorm"

Value

dataframe where fields with censored data (i.e., Surv objects) are transformed into a simple numeric fields

Details

The imputeOption values of "lower", "upper" and "mid" impute the lower limit, upper limit, and midpoint between the lower and upper limit. In the context of typical water quality data, these options would be equivalent to zero, detection limit and 1/2 detection limit substitution. Options for substituting the normal ["norm"] or lognormal ["lnorm"] expectation can also be used.

See also

Examples

if (FALSE) {
df  <- dataCensored[1:20, ]
df.lower <- imputeDF(df,'lower')
df.mid   <- imputeDF(df,'mid')
df.upper <- imputeDF(df,'upper')
df.norm  <- imputeDF(df,'norm')
df.lnorm <- imputeDF(df,'lnorm')
}