Impute Censored Values

Imputes substitute values for left-censored values using the expectation maximization method.

imputeLessThans(
  ...,
  type = c("MLE", "robust"),
  initial = c("complete.obs", "multRepl")
)

# S3 method for default
imputeLessThans(
  ...,
  type = c("MLE", "robust"),
  initial = c("complete.obs", "multRepl")
)

# S3 method for data.frame
imputeLessThans(
  ...,
  group = NULL,
  type = c("MLE", "robust"),
  initial = c("complete.obs", "multRepl")
)

Arguments

...
either a data frame that contains columns of class "qw" or any combination of individual vectors of class "numeric," "lcens," or "qw." Missing values are removed before processing.

type
the type of estimate, "MLE" for maximum likelihood estimates, or "robust" for robust estimation methods. See lrEM for details.

initial
the method to use for the initial log-ratio covariance matrix, either "complete.obs" that uses only the rows with no censored data to construct the matrix, or "multRepl" that uses simple substitution of censored values to compute the matrix. See lrEM for details.

group
character string, the name of the column in the data frame to indicate a group for imputation. See Details

Value

A data frame containing the original data with imputed censored values.

Details

Imputation of left-censored data requires the assumption of multivariate log-normality for a single population. If the data represent samples from multiple populations, then they should be identified by the group argument. The minimum size for any group is 3.

See also