Left-censored Data

Class "lcens" describes left-censored data. From smwrQW package.

Slots

.Data

a 2-column matrix that contains the value and the detection limit of the left-censored data. Missing values are permitted.

censor.codes

a logcial value, TRUE indicates a censored value, FALSE indates an uncensored value. Missing values are permitted only for missing values in .Data.

names

a character vector of unique names for each element in .Data.

Objects from the Class

Objects can be created by calls to as.lcens.

References

Lorenz, D.L., in preparation

Examples

showClass("lcens")

#> Class "lcens" [package "baytrends"]
#> 
#> Slots:
#>                                              
#> Name:         .Data censor.codes        names
#> Class:       matrix      logical    character
#> 
#> Extends: 
#> Class "matrix", from data part
#> Class "array", by class "matrix", distance 2
#> Class "mMatrix", by class "matrix", distance 2
#> Class "structure", by class "matrix", distance 3
#> Class "vector", by class "matrix", distance 4, with explicit coerce