Mutliply-censored Data

Class "mcens" describes right- or multiply-censored data. From smwrQW package.

Slots

.Data

a 2-column matrix that contains the lower and upper value for the data. Missing values are permitted.

censor.codes

an integer vector that indicates censoring; -1 indicates left-censored, 0 indicates uncensored, and 1 indicates rigth-censored. Missing values are permitted only for elements of .Data that are missing.

interval

a logical value; TRUE indicates that the value is interval censored and FALSE indicates otherwise. Missing values are permitted only for elements of .Data that are missing.

names

a character vector of unique names for each element in .Data.

Note

Note that censor.codes and interval are not necessary, but included to make some manipulations easier.

Objects from the Class

Objects can be created by calls to as.mcens.

References

Lorenz, D.L., in preparation

Examples

showClass("mcens")

#> Class "mcens" [package "baytrends"]
#> 
#> Slots:
#>                                                           
#> Name:         .Data censor.codes     interval        names
#> Class:       matrix      integer      logical    character
#> 
#> Extends: 
#> Class "matrix", from data part
#> Class "array", by class "matrix", distance 2
#> Class "mMatrix", by class "matrix", distance 2
#> Class "structure", by class "matrix", distance 3
#> Class "vector", by class "matrix", distance 4, with explicit coerce