Selected USGS parameter Codes

Selected U.S. Gelogical Survey (USGS) parameter codes to create column names and data type conversion when importing data from NWISweb.

From swmrQW

pcodeColData

Format

Data frame with 1178 rows and 3 columns

Name	Type	Description

parm_cd	character	The 5-digit parameter code

col_name	character	The column name to use when importing data from NWISweb

data_type	character	The data type to convert the imported data

Note

This data set can be copied and modifed by the user to add or change column names. The data_type must be either "qw" or "numeric" any other value will be ignored.

References

Lorenz, D.L., 2016, smwrQW---An R Package for Managing and Analyzing Water-Quality Data, Version 0.7.3

Examples

data(pcodeColData)
# just print some rows
head(pcodeColData)

#>     parm_cd            col_name data_type
#> 314   00001 LocationXSect_00001   numeric
#> 315   00002 LocationXSect_00002   numeric
#> 621   00003      SampleDepth_ft   numeric
#> 804   00004         StreamWidth   numeric
#> 316   00005 LocationXSect_00005   numeric
#> 503   00008        SampleAcctNo   numeric

tail(pcodeColData)

#>     parm_cd              col_name data_type
#> 720   99987   SolarRadiationTotal   numeric
#> 414   99988 Par(fdoahsdmimoppsmps   numeric
#> 416   99989             PAR_99989   numeric
#> 569   99993 SampleVolumeSchedOCSS   numeric
#> 610   99996     SampleWtSchedOCSS   numeric
#> 415   99997             PAR_99997   numeric