Parameter Code Information

Returns information about parameter codes. Support function for importNWISqw.

pcodeNWISqw(
  params = "all",
  group = TRUE,
  name = TRUE,
  CASRN = FALSE,
  short = TRUE,
  units = TRUE,
  col.name = FALSE
)

Arguments

params
A character string contains the name of a group of parameter codes, or a vector of pamater codes. See Details.

group
include the parameter group in the output?

name
include the parameter name in the output?

CASRN
include the parameter Chemical Abstracts Service Registry Number (CASRN) in the output?

short
include the parameter short name in the output?

units
include the parameter units in the output?

col.name
include a column name for the parameter in the output?

Value

A data frame with the parameter codes and the selected columns.

Details

Valid groups for param are "All," "INF," "PHY," "INM," "INN," "NUT," "MBI," "BIO," "IMN," "IMM," "TOX," "OPE," "OPC," "OOT," "RAD," "SED," or "POP."

Added from smwrQW.

References

Lorenz, D.L., 2014, smwrQW OFR.
See information about discrete samples at https://nwis.waterdata.usgs.gov/usa/nwis/qw.

See also

Examples

if (FALSE) {
pcodeNWISqw("00925")
}