Sample Quantiles

Computes sample or estimated quantiles corresponding to the given probabilities: methods for "lcens," "mcens," and "qw" data. The smallest observation (censored or uncensored) corresponds to a probability of 0 and the largest to a probability of 1.

# S3 method for lcens
quantile(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = TRUE,
  method = "flipped K-M",
  type = 2,
  alpha = 0.4,
  ...
)

# S3 method for mcens
quantile(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = TRUE,
  method = "flipped K-M",
  type = 2,
  alpha = 0.4,
  ...
)

# S3 method for qw
quantile(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = TRUE,
  method = "flipped K-M",
  type = 2,
  alpha = 0.4,
  ...
)

Arguments

x
an object of a censored-data class whose sample quantiles are wanted. NA and NaN values are not allowed unless na.rm is TRUE.

probs
numeric vector of probabilities with values in [0,1].

na.rm
logical; if TRUE, any NA and NaNs are removed from x before the quantiles are computed.

names
logical; if true, the result has a names attribute.

method
the method to use for computing quantiles. See Details.

type
an integer between 1 and 9 selecting one of the nine quantile algorithms described in quantile.

alpha
the offset fraction to be used, depending on method; typically in [0, 0.5].

...
not used, required for other methods.

Value

An optionally named vector of the requested probabilities. The names of values that would be left-censored are marked with "*."

Details

The methods available in the current version are "flipped K-M," "log ROS," "ROS," "log MLE," and "MLE." The method "flipped K-M" produces quantiles using the Kaplan-Meier method on flipped data described by Helsel (2012). The methods "log ROS" and "log MLE" are described by Helsel, 2012 and Helsel and Cohn (1988). The methods "ROS" and "MLE" are similar to "log ROS" and "log MLE" except that no log- and back-transforms are made on the data.

References

Helsel, D.R. 2012, Statistics for censored environmental data using Minitab and R: New York, Wiley, 324 p.
Helsel, D.R. and Cohn, T.A., 1988, Estimation of descriptive statistics for multiply censored water quality data: Water Resources Research v. 24, n. 12, pp. 1997-2004

See also

Examples

set.seed(28)
Xu <- rnorm(23)
quantile(as.lcens(Xu, 0))

#>       0%*      25%*      50%*       75%      100% 
#> 0.0000000 0.0000000 0.0000000 0.3840304 1.8414814