Water-quality Data

Class "qw" describes water-quality data. From smwrQW package.

Slots

.Data

a 2-column matrix that contains the lower and upper limit of the value of the water-quality data. Missing values are permitted.

remark.codes

a character vector of remark codes. Only "<" and ">" are meaningful for the class; other values will be ignored, but may generate warnings. Missing values are permitted only for corresponding missing values in .Data.

value.codes

a character vector for user-defined code that may help characterize the value of the water-quality data. Missing values are permitted only for corresponding missing values in .Data.

reporting.level

a numeric vector describing the reporting level of the water-quality data. Missing values are permitted if the reporting level is not known.

reporting.method

a character vector for user-defined code or description of the reporting.level. Generally used to distinguish between detection limits and quantitation levels.

reporting.units

a character vector describing the units. Some values, like "mg/L" are recognized by some applications.

analyte.method

a character vector to describe the analytical method.

analyte.name

a character vector of the analyte name. Some values may be recognized by some applications.

rounding

a vector of length 2 that describes how to round the data for printing. The first value is the number of significant digits, the second is the maximum number of decimal places, default is c(2,3).

unique.code

a character vector that defines a distinct collection of data. Typically used to define columns is datasets organized by sample.

names

a character vector of unique names for each element in .Data.

Objects from the Class

Objects can be created by calls to as.qw.

References

Lorenz, D.L., in preparation

Examples

showClass("qw")

#> Class "qw" [package "baytrends"]
#> 
#> Slots:
#>                                                                           
#> Name:             .Data     remark.codes      value.codes  reporting.level
#> Class:           matrix        character        character          numeric
#>                                                                           
#> Name:  reporting.method  reporting.units   analyte.method     analyte.name
#> Class:        character        character        character        character
#>                                                          
#> Name:          rounding      unique.code            names
#> Class:          numeric        character        character
#> 
#> Extends: 
#> Class "matrix", from data part
#> Class "array", by class "matrix", distance 2
#> Class "mMatrix", by class "matrix", distance 2
#> Class "structure", by class "matrix", distance 3
#> Class "vector", by class "matrix", distance 4, with explicit coerce