qw.import

Import the contents of tab delimited text files to a QW object.

qw.import(fn.import, qw.names, rounding = c(3, 4))

Arguments

fn.import
Filename (with path) containing data. Comma separated (CSV) format only.

qw.names
Names of parameters for QW columns.

rounding
OPTIONAL. Rounding values for QW columns. Default is c(3,4) but can be changed.

Value

Returns a QW enabled data frame.

Details

This internal function will import data from a comma delimited (CSV) file to create a data frame with columns of the qw class.

Only three slots are recognized by this function; ".Data" and "remark.codes" The slot name ".Data" includes ".Data.values" and ".Data.value2".

These three slots are marked as "_lo", "_hi", and "_symbol" in the import file the prefix of the parameter name to be used in the qw object; e.g., secchi_lo, secchi_hi, and secchi_symbol.

To view all qw slot names use the command slotNames("qw").

The function returns a data frame with QW enabled columns. The user must save the file. Any modification of column classes needs to be handled by the user (e.g., character to date or POSIXct).

Examples

# Use internal function to export dataCensored as example for import
qw.export(dataCensored, getwd(), "dataCensored_TEST.csv")

#> [1] "QW object saved to user named file in user specified directory;"
#> [1] "C:/Users/Erik.Leppo/OneDrive - Tetra Tech, Inc/MyDocs_OneDrive/GitHub/baytrends/docs/reference"

# Import Test file as a qw object

# Define function parameters
fn.import <- file.path(".", "dataCensored_TEST.csv")
qw.names <- c("secchi", "chla", "do", "tn", "tp", "po4", "tdp"
              , "no23", "nh4", "tdn", "tss")
rounding <- c(3, 4)

# Import
dataCensored.test<- qw.import(fn.import, qw.names, rounding)

#> [1] "Processing; 1/11 (secchi)."
#> [1] "Processing; 2/11 (chla)."
#> [1] "Processing; 3/11 (do)."
#> [1] "Processing; 4/11 (tn)."
#> [1] "Processing; 5/11 (tp)."
#> [1] "Processing; 6/11 (po4)."
#> [1] "Processing; 7/11 (tdp)."
#> [1] "Processing; 8/11 (no23)."
#> [1] "Processing; 9/11 (nh4)."
#> [1] "Processing; 10/11 (tdn)."
#> [1] "Processing; 11/11 (tss)."

# Check for qw class
str(dataCensored.test)

#> 'data.frame':	13062 obs. of  23 variables:
#>  $ station        : chr  "CB3.3C" "CB3.3C" "CB3.3C" "CB3.3C" ...
#>  $ date           : chr  "1985-05-21" "1985-05-21" "1985-05-21" "1985-05-21" ...
#>  $ layer          : chr  "AP" "B" "BP" "S" ...
#>  $ salinity_lo    : chr  "10.5" "14.4" "14.4" "10.105" ...
#>  $ salinity_hi    : chr  "10.5" "14.4" "14.4" "10.105" ...
#>  $ salinity_symbol: chr  " " " " " " " " ...
#>  $ wtemp_lo       : chr  "18.1" "15.3" "15.3" "18.8" ...
#>  $ wtemp_hi       : chr  "18.1" "15.3" "15.3" "18.8" ...
#>  $ wtemp_symbol   : chr  " " " " " " " " ...
#>  $ din_lo         : chr  "0.28" "0.372" "0.375" "0.259" ...
#>  $ din_hi         : chr  "0.28" "0.372" "0.375" "0.259" ...
#>  $ din_symbol     : chr  " " " " " " " " ...
#>  $ secchi         : qw, no censoring: "  NA" "  NA" ...
#>  $ chla           : qw, left censoring: "   7.000" "   3.000" ...
#>  $ do             : qw, no censoring: "  6.2" "  0.0" ...
#>  $ tn             : qw, no censoring: " 0.651" " 0.727" ...
#>  $ tp             : qw, no censoring: " 0.0512" " 0.0310" ...
#>  $ po4            : qw, left censoring: " 0.0030" " 0.0063" ...
#>  $ tdp            : qw, left censoring: " 0.0360" " 0.0170" ...
#>  $ no23           : qw, no censoring: " 0.1760" " 0.1170" ...
#>  $ nh4            : qw, left censoring: " 0.1040" " 0.2550" ...
#>  $ tdn            : qw, no censoring: "  0.5200" "  0.6400" ...
#>  $ tss            : qw, left censoring: "   6.90" "   7.60" ...

# Save
save(dataCensored.test, file="dataCensored.test.rda")

# Show slot names for a qw object.
slotNames("qw")

#>  [1] ".Data"            "remark.codes"     "value.codes"      "reporting.level" 
#>  [5] "reporting.method" "reporting.units"  "analyte.method"   "analyte.name"    
#>  [9] "rounding"         "unique.code"      "names"           

####
# convert date field to POSIXct
#dataCensored.test[,"date"] <- as.POSIXct(dataCensored.test[,"date"])
# str(dataCensored.test)
# as.numeric() and as.integer() can be used to convert columns of those types.