Extract Model Residuals

Extracts the residuals from a censored regression object: method for "censReg" object.

# S3 method for censReg
residuals(object, type = "working", suppress.na.action = FALSE, ...)

Arguments

object
an object of class "censReg"---output from censReg

type
The type of residuals, see Details.

suppress.na.action
logical, suppress the effects of the na.action in the call to censReg and return only the fitted values corresponding to the fitted data.

...
further arguments passed to or from other methods.

Value

The residuals from the regression as specified by type.

Details

The value for type can be any one of the following:

Value	Description

"working"	Residuals with censored residuals replaced by their expected values

"response"	Residuals from the linear predictor

"influence"	An estimate of Cook's D values based on "working" residuals

"leverage"	The hat diagonals

"S-L"	The square-root of the absolute value of the residuals with censored residuals replaced by their expected value

Also, any other value of type for residuals.survreg can be used to obtain those residuals. Note that "working" and "response" are defined in the table above, in keeping with older versions of censReg.