Explanation

Adds the current plot information to plot control list (support function).

setExplan(current, old = NULL)

Arguments

current
the current plot information.

old
existing explanation information.

Value

A list having four components:

text

a list having two components:

ll text the description in the explanation, derived from the name component in Plot
cex the size of the text to write in the explanation

lines

a list having seven components and controlling both lines and points:

ll type the type of plot (specifying points, lines, and so forth)
lwd the line weight
lty the line type
pch the symbol
csi the size of the symbol in inches
cex the the size of the symbol relative to character size
col the color of the plot

areas

a list having two components:

ll fill the color of the fill area
borders the color of the border

current

a list like current with the defaults set

Each entry in text must have a corresponding entry in lines and areas.
Details

Added from smwrGraphs.