Plot Parameters

Sets the plot control list (support function).

setPlot(current, name = "", what = "lines", type = "solid",
  width = "standard", symbol = "circle", filled = TRUE,
  size = 0.09, color = "black", area.color = NA, area.border = NA)

Arguments

current
list containing the current plot information or those requested by the user.

name
the name of the object plotted; used in the explanation. Expressions can be used, but if used, then name must be and expression for all calls that update the current plot information.

what
what to plot, see Details.

type
the line type, if drawn, must be one of "solid," "dashed," or "dotted."

width
the width of the line, if drawn, must be one of "standard;" "color," a little wider than "standard;" "bold," substantially wider than "standard;" or "hairline," used for ticks and borders.

symbol
the symbol to plot, if drawn, see Details.

filled
logical, if TRUE and a symbol is drawn, then fill with solid color.

size
the size of the symbol, in inches, if drawn.

color
the color of the symbol or line.

area.color
the color of a shaded area, required for completeness.

area.border
the boundary color of a shaded area, required for completeness.

Value

A list like current, but with the defaults supplied for any missing component.

Details

Added from smwrGraphs.

The value for what must be one of:

"points"

symbols only,

"lines"

lines only,

"both"

lines connecting symbols with a small gap,

"overlaid"

lines connecting symbols,

"stairstep"

horizontal line to next x value with a vertical line to the y value,

"vertical"

vertical lines from y equal 0 only.

"none"

draw nothing

The value for symbol must be one of:

"circle"

an open or filled circle, depending on filled,

"uptri"

an open or filled up pointing triangle, depending on filled,

"plus"

a plus sign (never filled),

"x"

an x (never filled),

"diamond"

an open or filled diamond shape, depending on filled,

"downtri"

an open or filled down pointing triangle, depending on filled,

"square"

an open or filled square, depending on filled,

"dot"

a very small dot (never filled),

"+"

a plus sign (never filled),

"none"

no symbol or line.

Note

Vertical lines drawn by setting what to "vertical" are drawn from y equal 0 to each y value. The user will have more control over vertical lines by using the addBars function and setting the bar width to 0.